Move ConnectionEvent code out of interface
