+++ examples/udp_server.cpp
@@ -42,10 +42,10 @@ int main(int argc, char *argv[])
 	srvr->startServer();
 	std::cout << "Server started\n";
 
-	auto connectionEventHandler = [](int clientId, sockep::IServerSockEP::ConnectionEvent event)
+	auto connectionEventHandler = [](int clientId, sockep::ConnectionEvent event)
 	{
 		std::cout << "Client " << clientId << " (" << srvr->getClientAddress(clientId) << ") "
-		          << srvr->getConnectionEventName(event) << std::endl;
+		          << stringifyConnectionEvent(event) << std::endl;
 	};
 	srvr->registerConnectionEventHandler("main", connectionEventHandler);
