+++ include/sockep/DataTypes.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <string>
+
+namespace sockep
+{
+
+enum ConnectionEvent
+{
+	CONNECTED,
+	DISCONNECTED
+};
+
+std::string stringifyConnectionEvent(ConnectionEvent event);
+
+} // namespace sockep
